vulkan: sub-step 1 is built and tested; record what the previous commit swept in
The Vulkan backend's storage and submission layer, its test and the CMake
changes were committed in 757dee9, the fork gate record, by an unscoped
add before they had been built. That commit's message says nothing about
them. This one does: the default tree was rebuilt and is unchanged since
the option is off, the option-to-flag conversion the generated config
needed is exercised by every configure, and the Vulkan tree was built
against the installed SDK and its test passes 13 checks on the Radeon VII.
The fork gate record also had the wrong system load for its first two
cells; corrected from the archived monitors.

+++ tests/backend_vulkan.cpp
@@ -56,7 +56,7 @@ int main() {
         std::vector<uint8_t> src = pattern(3 * mib + 12345, 1);
         const std::vector<uint8_t> kept = src;
         const auto adopted = b->adopt(src.data(), src.size());
-        std::fill(src.begin(), src.end(), 0);
+        std::fill(src.begin(), src.end(), uint8_t(0));
         out.assign(kept.size(), 0);
         b->read(*adopted, 0, out.data(), kept.size());
         require(out == kept, "adopted bytes differ after upload");
